requirements.txt
Django>=4.0,<5.0
gunicorn
whitenoise
djangorestframework
djangorestframework-simplejwt
django-cors-headers
